traj_action: move is_*(..) to utils
